# openapi.yaml
openapi: 3.1.0
info:
  title: GSTD Sovereign AI Platform API
  description: |
    Decentralized AI inference, federated learning, and compute marketplace on TON blockchain.
    
    **Quick Start:**
    1. Set Base URL: `https://api.gstdtoken.com/v1`
    2. Generate API Key in the Dashboard (Sovereign Switch → API Gateway)
    3. Use `Authorization: Bearer gstd_...` header
    
    Compatible with OpenAI SDK, Cursor, VS Code, LangChain, and any tool supporting OpenAI API format.
  version: 2.0.0
  contact:
    name: GSTD Platform
    url: https://app.gstdtoken.com
  license:
    name: MIT
servers:
  - url: https://api.gstdtoken.com/v1
    description: Production Gateway
  - url: https://app.gstdtoken.com/api/v1
    description: Platform API
security:
  - BearerAuth: []
  - SessionToken: []
paths:
...
  /pipeline/register:
    post:
      tags: [Infrastructure]
      summary: Register GPU Node for Pipeline
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id: { type: string }
                vram_mb: { type: integer }
                gpu_model: { type: string }
                region: { type: string }
      responses:
        '200': { description: Node registered }

  # ====== MOBILE COMPUTE ======
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "GSTD API Key (gstd_...)"
    SessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: "Session token from wallet login"
...